refactor: apply refactoring to rebased code
