+++ packages/apps/reputation-oracle/server/src/integrations/slack-bot-app/slack-bot-app.ts
@@ -3,7 +3,7 @@ import { View, ViewsOpenResponse } from '@slack/web-api';
 import { IncomingWebhookSendArguments } from '@slack/webhook';
 import { firstValueFrom } from 'rxjs';
 import logger from '../../logger';
-import { formatAxiosError } from '../../utils/format-axios-error';
+import * as httpUtils from '../../utils/http';
 
 export class SlackBotApp {
   private readonly logger = logger.child({ context: SlackBotApp.name });
@@ -18,7 +18,7 @@ export class SlackBotApp {
         this.httpService.post(this.config.webhookUrl, message),
       );
     } catch (error) {
-      const formattedError = formatAxiosError(error);
+      const formattedError = httpUtils.formatAxiosError(error);
       const errorMessage = 'Error sending Slack notification';
       this.logger.error(errorMessage, {
         error: formattedError,
@@ -50,7 +50,7 @@ export class SlackBotApp {
         throw new Error('Error opening Slack modal');
       }
     } catch (error) {
-      const formattedError = formatAxiosError(error);
+      const formattedError = httpUtils.formatAxiosError(error);
       const errorMessage = 'Error opening Slack modal';
       this.logger.error(errorMessage, {
         error: formattedError,
@@ -63,7 +63,7 @@ export class SlackBotApp {
     try {
       await firstValueFrom(this.httpService.post(responseUrl, { text }));
     } catch (error) {
-      const formattedError = formatAxiosError(error);
+      const formattedError = httpUtils.formatAxiosError(error);
       const errorMessage = 'Error updating Slack message';
       this.logger.error(errorMessage, {
         error: formattedError,
